↔️ Split Up randomBetween() and pickitt(). Make index.js only export.
Issues #39, #40
