+++ package.json
@@ -8,7 +8,7 @@
     "package.json",
     "README.md",
     "index.js",
-    "index.ts"
+    "src"
   ],
   "scripts": {
     "test": "jest --coverage",
